Refactor TestDE for S4 objects rather than lists, use ggplot2.
The objects are nested, with the outer layer "BASiCS_ResultsDE"
being basically a wrapper for "BASiCS_ResultDE", which hold
the result for each test. This can be further improved by
adding some helper functions for the outer and inner classes.

It's questionable how much of the objects need to be stored.
Some of it could be calculated on the fly, to reduce the
complexity of the classes. Also, adding functionality
to annotate genes with rowData/metadata would be
very nice. Subset functions (to remove/retain genes)
would also be very nice but have not yet been added.

This is a massively breaking change. Could consider adding
functions to convert from old to new and new to old
temporarily.